style: filter styling phase 2
