+++ src/components/features/products-listing/ProductsListing.css
@@ -1,4 +1,44 @@
-.products-listing-section{
-    display: flex;
-    flex-direction: column;
+.products-listing-section {
+  display: flex;
+  flex-direction: column;
+}
+
+.toggle-filter-section {
+  width: 100%;
+  background: whitesmoke;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.4);
+  padding: 1rem;
+  display: flex;
+  justify-content: space-between;
+}
+
+.toggle-filter-btn {
+  cursor: pointer;
+  display: inline-block;
+}
+
+.products-listing {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+  align-items: center;
+  padding: 1rem;
+  gap: 1rem;
+}
+
+.product-contents {
+  border: 1px solid;
+  height: 300px;
+  width: 200px;
+}
+
+@media (min-width: 768px) {
+  .toggle-filter-wrapper {
+    display: none;
+  }
+
+  .filters-products-layout {
+    display: grid;
+    grid-template-columns: 16rem auto;
+  }
 }
