+++ src/components/features/products-listing/ProductsListing.jsx
@@ -1,10 +1,14 @@
 import { Link, useNavigate } from "react-router-dom";
+import { useState } from "react";
 
+import "./ProductsListing.css";
 import { useProducts } from "../../../core/contexts/products-context/ProductsContext";
 import { useAuthentication } from "../../../core/contexts/authentication-context/AuthenticationContext";
 import Filters from "./filters/Filters";
+import { MdClose, MdFilterList } from "react-icons/md";
 
 const ProductsListing = () => {
+  const [toggleFilterSection, setToggleFilterSection] = useState(false);
   const navigate = useNavigate();
   const { token } = useAuthentication();
 
@@ -22,61 +26,71 @@ const ProductsListing = () => {
   const isWishlistContainsProduct = (productId) =>
     wishlist.find((product) => product._id === productId);
 
+  const filterHamburgerHandler = () => {
+    setToggleFilterSection((prev) => !prev);
+  };
+
   return (
     <div className="products-listing-section">
-      <Filters />
-      <div
-        style={{
-          display: "flex",
-          justifyContent: "center",
-        }}
-      >
-        {products.length > 0 &&
-          products.map((product) => {
-            const { _id, title, author, price, categoryName, rating } = product;
-            return (
-              <div
-                key={_id}
-                style={{
-                  border: "1px solid",
-                  height: "300px",
-                  width: "200px",
-                }}
-              >
-                <h3>{title}</h3>
-                <p>{author}</p>
-                <p>{price}</p>
-                <p>{categoryName}</p>
-                <p>{rating}⭐</p>
-                <Link to={`/product/${_id}`}>View details</Link>
-                <button
-                  onClick={() =>
-                    token
-                      ? isCartContainsProduct(_id)
-                        ? navigate("/cart")
-                        : addProductToCart(product)
-                      : navigate("/login")
-                  }
-                >
-                  {isCartContainsProduct(_id) ? "Go to Cart" : "Add to Cart"}
-                </button>
-                <button
-                  onClick={() =>
-                    token
-                      ? isWishlistContainsProduct(_id)
-                        ? removeProductFromWishlist(_id)
-                        : addProductToWishlist(product)
-                      : navigate("/login")
-                  }
-                >
-                  {isWishlistContainsProduct(_id)
-                    ? "Remove from Wishlist"
-                    : "Add to wishlist"}
-                </button>
-              </div>
-            );
-          })}
-        {products?.length < 1 && <h3>No products found.</h3>}
+      <div className="toggle-filter-section">
+        <div className="toggle-filter-wrapper">
+          <div className="toggle-filter-btn">
+            {toggleFilterSection ? (
+              <MdClose size={15} onClick={filterHamburgerHandler} />
+            ) : (
+              <MdFilterList size={15} onClick={filterHamburgerHandler} />
+            )}
+          </div>
+          <span>{toggleFilterSection ? "Close" : "Apply Filters"}</span>
+        </div>
+        {products.length > 0 && <p>{products.length} results</p>}
+      </div>
+      <div className="filters-products-layout">
+        <Filters
+          className={toggleFilterSection ? "" : "hide-filters-wrapper"}
+        />
+        <div className="products-listing">
+          {products.length > 0 &&
+            products.map((product) => {
+              const { _id, title, author, price, categoryName, rating } =
+                product;
+              return (
+                <div key={_id} className="product-contents">
+                  <h3>{title}</h3>
+                  <p>{author}</p>
+                  <p>{price}</p>
+                  <p>{categoryName}</p>
+                  <p>{rating}⭐</p>
+                  <Link to={`/product/${_id}`}>View details</Link>
+                  <button
+                    onClick={() =>
+                      token
+                        ? isCartContainsProduct(_id)
+                          ? navigate("/cart")
+                          : addProductToCart(product)
+                        : navigate("/login")
+                    }
+                  >
+                    {isCartContainsProduct(_id) ? "Go to Cart" : "Add to Cart"}
+                  </button>
+                  <button
+                    onClick={() =>
+                      token
+                        ? isWishlistContainsProduct(_id)
+                          ? removeProductFromWishlist(_id)
+                          : addProductToWishlist(product)
+                        : navigate("/login")
+                    }
+                  >
+                    {isWishlistContainsProduct(_id)
+                      ? "Remove from Wishlist"
+                      : "Add to wishlist"}
+                  </button>
+                </div>
+              );
+            })}
+          {products?.length < 1 && <h3>No products found.</h3>}
+        </div>
       </div>
     </div>
   );
